+++ Models/Member.cs
@@ -68,7 +68,7 @@ namespace PinewoodGrow.Models
 		[Required(ErrorMessage = "You cannot leave the Income blank.")]
 		[DataType(DataType.Currency)]
 		[Range(0.0, 22186, ErrorMessage = "Income must be between $0 and $22,186.")]
-		public decimal Income { get; set; }
+		public double Income { get; set; }
 
 		[StringLength(2000, ErrorMessage = "Member Notes cannot be more than 2000 characters long.")]
 		[DataType(DataType.MultilineText)]
